# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GuardRail API
  description: >-
    GuardRail — Privileged Access Management for browser-based access to web
    management interfaces. All endpoints are under `/api/v1`. Errors use RFC 9457
    `application/problem+json`. Authentication is a Bearer access token (JWT);
    the refresh token is delivered as an HttpOnly cookie.
  version: 1.0.0
  license:
    name: Proprietary
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: List devices
      parameters:
        - { in: query, name: vendor, schema: { type: string } }
        - { in: query, name: tag, schema: { type: string } }
        - { in: query, name: q, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Device" } }
    post:
      tags: [Devices]
      summary: Register a device
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
  /devices/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Devices]
      summary: Get a device
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [Devices]
      summary: Update a device
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
    delete:
      tags: [Devices]
      summary: Soft-delete a device
      responses:
        "204": { description: No Content }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    DeviceInput:
      type: object
      required: [name, host]
      properties:
        name: { type: string }
        description: { type: string }
        vendor: { type: string }
        device_type: { type: string }
        host: { type: string }
        port: { type: integer }
        scheme: { type: string, enum: [http, https] }
        verify_tls: { type: boolean }
        custom_headers: { type: object, additionalProperties: { type: string } }
        tags: { type: array, items: { type: string } }
        requires_approval: { type: boolean }
    Device:
      allOf:
        - $ref: "#/components/schemas/DeviceInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            status: { type: string }
            url: { type: string }
...